Feedback management microservice refactor (#1057)
