+++ comps/feedback_management/deployment/docker_compose/compose_feedback_mongo.yaml
