+++ comps/feedback_management/src/Dockerfile
@@ -20,11 +20,11 @@ COPY comps /home/user/comps
 COPY requirements.txt /home/user/
 
 RUN pip install --no-cache-dir --upgrade pip setuptools && \
-    pip install --no-cache-dir -r /home/user/comps/feedback_management/mongo/requirements.txt && \
+    pip install --no-cache-dir -r /home/user/comps/feedback_management/src/requirements.txt && \
     pip install --no-cache-dir -r /home/user/requirements.txt
 
 ENV PYTHONPATH=$PYTHONPATH:/home/user
 
-WORKDIR /home/user/comps/feedback_management/mongo
+WORKDIR /home/user/comps/feedback_management/src
 
-ENTRYPOINT ["python", "feedback.py"]
+ENTRYPOINT ["python", "opea_feedback_microservice.py"]
