+++ comps/feedback_management/src/integrations/mongo/mongo_conn.py
@@ -4,7 +4,8 @@
 from typing import Any
 
 import motor.motor_asyncio as motor
-from config import DB_NAME, MONGO_HOST, MONGO_PORT
+
+from .config import DB_NAME, MONGO_HOST, MONGO_PORT
 
 
 class MongoClient:
