+++ comps/feedback_management/src/requirements.txt
